requirements.txt
# requirements.txt
python-dotenv
discord.py
discord-webhook